backend/requirements.txt
fastapi
uvicorn
web3
requests
pydantic
pydantic-settings
google-genai
